(PUP-9481) Add tests for different types of settings hooks
If a setting has a default value, then the value will be used, but it's hook
will only be called if it is of type `on_define_and_write`.

If a setting is defined in the `main` section, then the value will be used (if
not overridden in a different section) and its hook will be called.

If a setting is set in a section other than `main` and it doesn't match the
current application run mode or enviroment, then the setting's value will be
ignored and its hook will never be called.

If a setting is set in a section other than `main` and it does match the current
application run mode or environment, then the setting's value will be used, but
the hook will only be called if its type is `:on_initialize_and_write`.

As a result, if a setting is overridden, such as in `agent` and the hook is of
type :on_write_only (the default) or :on_define_and_write, then the hook will
never be called, which can be surprising if the setting performs validation in
its hook.