(PUP-10722) Fix masterport not honored
PUP-10670 introduced the `serverport` alias for `masterport` and changed
Puppet to use it internally instead of `masterport`.
This introduced the following change in behaviour:
- when setting `masterport` in any section different than `main`(eg.
  agent), this is ignored and the default `serverport` value is used.
- users that have `masterport` set in their configuration, in a
  different section than `main`(eg. agent), are also affected as Puppet
  will not pick up the value set for `masterport` and it will use the
  default `serverport` value.

This is due to the setting hooks not being called when they are set in
sections other than `main`.

In addition to the setting hooks, check and set masterport and
serverport when `initialize_app_defaults` is called, if required.