(PUP-10713) Pass expiration time to expired? method
Each entry defines `expires` and `expired?` methods, which basically do the same
thing. One is called by `clear_all_expired`, the other is called by
`evict_if_expired`. This commits switches to using `expired?` so that `expired`
can be deleted in a later commit.

As a result of this change, the `clear_all_expired` method compares each entry's
expiration time against the same time instance. So the method is no longer
sensitive to the amount of time it takes to evict an environment.