(PUP-10713) Recompute next_expiration time
The next_expiration time was only recomputed when `clear_all_expired` as called,
but not `get_conf`. Move that to `evict_if_expired` so it happens for both.
Puppet uses that variable to skip expiration checking if the next expiration
time is in the future and none of the environments are marked as expired in
puppetserver.

NEEDS TEST