(PUP-10713) Recompute next_expiration time
Previously the `next_expiration` time was recalculated when `Cached#get` cleared
all expired environments, but not when `Cached#get_conf` cleared the environment
if it was expired.

Move the recalculation to `evict_if_expired` so it happens for both code paths.

Tests aren't added because the `next_expiration` time will be removed in a
future commit.