(PUP-10713) Remove next expiration "optimization"
Previously we stored the expiration times for each environment in a
`expirations` sorted set and `next_expiration` was set to the first expiration
time in the set or END_OF_TIME if it was empty.

However, in order to determine if any environments need to be evicted, we have
to call the cache expiration service (in puppetserver) for all environments at
least once.

In the best case, either Time.now < next_expiration and no environments have
been evicted by puppetserver, or Time.now >= next_expiration and we still query
puppetserver for all environments.

In the worst case, Time.now < next_expiration and the last environment in the
cache is expired. In that case, we call the cache expiration service for each
environment again(!).

Since we have to query the cache expiration service at least once for each
environment, we might as well just ask it, and remove the `next_expiration`
"optimization". Doing so eliminates the need for the `Entry#expires` method.