(maint) [WIP] call :on_write_only on application initialization
Currently if a setting has a hook of type `:on_write_only`, it will be
called only when it's set in the `[main]` or `[user]` section.

Assuming we have the following configuration where all mentioned
settings have hooks of type `:on_write_only`.

```
[main]
main_setting = a

[agent]
agent_setting = b
```

Only the hook for `main_setting` will be called, regardless of our run
mode (it doesn't matter what puppet application we're running).

This behavior is due to puppet collecting the hooks before our
application is initialized[1]. As settings are collected based on the
current run mode (which is always `:user` before application
initialization), settings set in sections other than `[user]` or
`[main]` will always be ignored.

The issue only manifests at the beginning of an application (i.e. when
the config file is parsed), so it's not an issue if a setting is written
to programmatically (such as `Puppet[:agent_setting] = 'c'`).

A suggested naive fix would be to--after the current application is
initialized and the run mode is switched--collect the settings again
(this time with the correct run mode) and evaluate hooks.

We don't care about the other hook types, as they are handled separately
in [2], [3] (2 times here due to the `#initialize_app_defaults` being
called twice: first in a global context, and then in the context of our
current application), and [4].

I don't think this change on its own obsoletes the work in #8411, but
it fixes the issue of `:on_write_only` hooks not behaving consistently
across sections.

Also, this doesn't fix the issue of `puppet config print agent_setting
--section agent` not honoring the hook of `agent_setting`.

[1] https://github.com/puppetlabs/puppet/blob/e06b87a71b09e7598182c9753f6151bd4155e659/lib/puppet/settings.rb#L609
[2] https://github.com/puppetlabs/puppet/blob/e06b87a71b09e7598182c9753f6151bd4155e659/lib/puppet/settings.rb#L632
[3] https://github.com/puppetlabs/puppet/blob/e06b87a71b09e7598182c9753f6151bd4155e659/lib/puppet/settings.rb#L386
[4] https://github.com/puppetlabs/puppet/blob/e06b87a71b09e7598182c9753f6151bd4155e659/lib/puppet/settings.rb#L980