(PUP-10730) Convert puppet http responses in rest termini
This is similar to what was done for the connection adapter, but this is for the
`parse_response` and `convert_http_http_error` methods used by rest termini.
Since the latter method now receives an instance of Puppet::HTTP::Response, we
call the `:reason` reader instead of `:message` and call `response.success?`.