(PUP-10772) Only accept msgpack if gem is installed
Previously the agent always advertised that it accepted the rich data msgpack
format (application/vnd.puppet.rich+msgpack) even if the agent didn't have the
msgpack gem installed locally. If the msgpack gem was installed on the server,
and the server selected the rich data msgpack format, then the agent would fail
to deserialize the catalog.

Now we call the `suitable?` method in the superclass so the msgpack feature
constrain is taken into account. Calling the super `supported?` method triggers
a bunch of other logic around whether or not the format implements required
methods, so avoid that.

Note the "raw" msgpack format didn't have this issue because it didn't override
the "supported?" method.