(maint) Make sure test classes implement the interface
Implement methods for test classes so that we can enable partial double
verification later. For example, we won't be able to add an expectation for
`AgentTestClient#run` if the class doesn't implement the method. It also makes
it less likely that our test will pass because the subject is overly stubbed,
such as trying to use a `notify` resource to test the relationship graph, yet
it doesn't implement `generate`.