(PUP-10778) Update regex for aix password parsing
Prior to this commit, the aix user resource only allowed for a static
pattern when parsing the passwd file for passwords. This would cause the
resource to be continually applied if the password line had spacing that
did not align with the pattern. This commit changes the parsing to allow
for abitrary spacing in the password line.