(PUP-10774) Do not query all groups/users
`get_posix_value` method checks if the user/group
that is added in the manifest corresponds with the
user/group retrieved from system. If they differ,
a full query is executed to retrieve all the
users/groups and match them with the required value.

This commit adds an extra check, for the case when
multiple users/groups have a same identifier(uid/gid)
In this case a full query is not needed, and the found
identifier is returned.