(PUP-10786) Update puppet node clean for ca-cli 2
Previously, the puppet node clean action called the PuppetServer CA CLI
as a library to clean certs. To do so it implemented a bridge between
the CA library's logger and Puppet's logger. That implementation only
included those methods used by the CA's clean action. However, in the
most recent version of the CA it will now also `warn` if the cadir is in
the legacy location (like Puppet proper does).

This patch implements the remaining API for the CA's logger (debug
and warn). It also filters out the warning from the CA library for a
cadir nested w/in the ssldir, as Puppet should have already warned about
that if needed.