(PUP-10797) Add runtime dependency on scanf
Previously, puppet had a dependency on the ruby `scanf` gem when running
puppet from source, but not when running as a gem. To ensure transitive
dependencies are resolved correctly, like when running puppetlabs-puppetdb
specs against puppet as a gem, move the `scanf` dependency to
`project_data.yaml`.

Note the runtime dependency is only needed when running ruby 2.7 or up. If
running on 2.5 or 2.6, then the `scanf` gem will be ignored, so there's no
harm in including it.