(PUP-10790) Fix regex warnings introduced in Ruby 2.7
Ruby 2.7 introduced a warning for checking invalid objects (such as
Integer in our case) against a regular expression. The check in this
case would always return nil even though a valid String would respect
the given regular expression, resulting in the possibility of unwanted
behaviour.

This commit replaces the check accordingly in the user type and keeps
previous logic.