(maint) Fix parallel:spec, 0 failures but exit code 1
Running the `parallel:spec` rake task on Windows with Ruby v2.5 started
to surface a constant issue in our tests: passing with 0 failures but
always completing with exit code 1. The `should raise an error if the
state file does not contain valid YAML and cannot be renamed` test from
`storage_spec.rb` was forcing the `File.rename` method to raise a
`SystemCallError` any time it was called as it didn't mention any
specific arguments. Since this is a fairly common method, this, at some
point, started being called more than once, before and apart from the
case that the test was expecting, causing the code to quit unexpectedly
and kill the parent thread of the parallel group that was currently
running.

This commit fixes the test by raising the needed
`SystemCallError` when specific arguments are found and allowing the
original `File.rename` method to be called. More information can be
found in the pull request's description.