(PUP-10818) Cache the deferred function, not its result
Merge commit a17761d60b introduced a semantic conflict between the environment
convergence changes in 5.5.x and the deferred function support in 6.x. As a
result puppet evaluated deferred functions before caching the result. This
regression was first introduced in 6.12.0. As a result, if the agent ran
from a cached catalog, it would use the result from last time, instead of
re-evaluating the deferred function.

This commit switches the order of those operations and adds a test. We have
to use `binary_file` because `file` is a 3x function and there are issues
with 3x functions and deferred values.