(maint) Allow additional respond_to arguments
Respond_to?(:getgrouplist) is called to determine if the host successfully
loaded the getgrouplist function from libc. It may be called with additional
arguments, so ignore them.

If respond_to?(:getgrouplist) returns false, then a LoadError is raised and
rescued, and the error message is logged at debug:

    Falling back to Puppet::Etc.group: The 'getgrouplist' method is not available

Remove extra whitespace.