(PUP-10819) Define and call 3x functions using the same Module object
Previously, puppet apply could call deferred 3x functions, but puppet agent
could not, resulting in:

    Function <name> not defined despite being loaded!

This was due to `puppet agent` calling `Puppet::Parser::Functions.environment_module`
with different `Puppet::Node::Environment` objects, but the same name. As a
result, 3x functions were defined on one Module instance[1], but a different
Module instance was mixed into the scope[2]. As a result, calling
`Scope#function_<name>` would trigger `method_missing`.

The reason this affected `puppet agent` is because it creates a
`Puppet::Node::Environment.remote` environment_instance when deserializing the
catalog[3], which is a different instance than the current environment pushed onto
the context.

This commit passes the current environment to the DeferredResolver rather than
using the catalog's `environment_instance`.

[1] https://github.com/puppetlabs/puppet/blob/7.1.0/lib/puppet/parser/functions.rb#L207
[2] https://github.com/puppetlabs/puppet/blob/7.1.0/lib/puppet/parser/scope.rb#L1124
[3] https://github.com/puppetlabs/puppet/blob/7.1.0/lib/puppet/resource/catalog.rb#L419