(PUP-9176) Requested module breaks existing module dependency
In 9822a26bf we only handled the case where a module that is to be
installed cannot satisfy one or more of its dependencies.

Spec tests uncovered a case that we missed, where the module satisfies
all of its constraints, but an existing installed module would become
unsatisfied after the installation.

I think this has a low chance of occuring in practice. To reproduce
this, one has to:

1. Install a module with fixed upper dependencies:
> puppet module install --target-dir tmpdir trepasi-debnet -v 1.5.1
This version has puppetlabs-stdlib bound to `>= 3.2.0 < 5.0.0`

2. Forcibly remove the `stdlib` module
> rm -rf tmpdir/stdlib

3. Attempt to install a version of stdlib that breaks the above-stated
   dependency
> puppet module install --target-dir tmpdir puppetlabs-stdlib -v "> 5"

Without this PR, the command fails with a cryptic NilClass error, due to
assumptions we made in the original fix for this ticket.

This commit handles this type of installation failure, and prints out a
list of all installed modules that depend on the module we're
installing, along with the range constraints.

The error message then becomes:

```
Error: Could not install module 'puppetlabs-stdlib' (> 5)
  The requested version cannot satisfy one or more of the following installed modules:
    puppetlabs-concat, expects 'puppetlabs-stdlib': >= 4.2.0 < 5.0.0
    trepasi-debnet, expects 'puppetlabs-stdlib': >= 3.2.0 < 5.0.0

  Use `puppet module install 'puppetlabs-stdlib' --ignore-dependencies` to install only this module
```