(PUP-10818) Validate catalog before caching it
Commit 408bf4e0f9 cached the resource catalog before resolving deferred
functions and converting the resource catalog to a RAL catalog. This way the
deferred function is cached, not the function's result. However, that caused
a regression as we cached the resource catalog before type/provider validation,
which occurs during `Catalog#to_ral`. For example, when using a non-absolute
path as the `command` in an `exec` resource.

Due to PUP-9323, we have to resolve deferred values, which mutates the resource
catalog, before converting to RAL and provider validation, and we need to do
that before caching. So make a copy of the original resource catalog, and cache
that if validation succeeds.