(PUP-10851) Omit error when parsing 'allow *' in fileserver.conf
Legacy auth was removed in 7, so if we parse an authz rule in fileserver.conf,
then we want to alert the user that it will be ignored by puppet. However,
`allow *` rules are harmless and match puppet's current behavior, so ignore
those, but continue logging an error for other allow/deny rules.