(PUP-10842) Fix production environment folder creation
Before this commit puppet expected the default `production` environment
folder to always be in the first path of the `environmentpath` setting
and was being automatically created if not found there.

This commit adds a check to make sure it was not already created in any
of the other environment paths given before automatically creating a new
one. This can happen if a new path is inserted at the beginning of the
`environmentpath` setting after the `production` folder was already
created. This change also aligns with custom environment behaviour where
it searches in the given environment paths and uses the first one it
matches, from left to right, the environment name given.