(PUP-10844) Continue when server_list server fails
When a server from server_list is available at the beginning of the run,
it will be cached, and the cached server will be used even if it is no
longer reachable.

Now, when a server from server_list becomes unavailable, the cached
server is invalidated and a new reachable server from server_list will
be selected.