(PUP-1357) Only skip a class scope the previous instance wasn't a node scope
Previously, puppet did not include a class if the class name happened to be the
same as the enclosing node definition, such as:

     node abc { include abc }

This occurred because puppet added the node scope to the `Scope#class_scope`
hash while evaluating the node definintion[1]. Later, puppet checked to see if a
class scope with the same name had already been added, and if so, puppet ignored
it[2]. However, that logic didn't account for the presence of a node scope with
the same name.

This commit causes puppet to exclude the node scope from its `already_excluded`
resources, so that it doesn't skip the class with the same name. It preserves
the behavior of only including a class once as implemented in[3].

[1] https://github.com/puppetlabs/puppet/blob/6.20.0/lib/puppet/resource/type.rb#L339
[2] https://github.com/puppetlabs/puppet/blob/6.20.0/lib/puppet/parser/compiler.rb#L473-L478
[3] https://github.com/puppetlabs/puppet/commit/927dff41df8f1c236c54eaee9fa1db7a3efaf02af