(maint) Correct value of SERVICE_CONFIG_LAUNCH_PROTECTED
SERVICE_CONFIG_LAUNCH_PROTECTED was given the value of 0x00000012 but that is
incorrect. The value is 12 decimal, which is 0x0C in hexadecimal, not 0x12.
0x12 is actually 18, which is not a documented ChangeServiceConfig2W
information level.

You can confirm this by checking the Microsoft documentation:

https://docs.microsoft.com/en-us/windows/win32/api/winsvc/nf-winsvc-changeserviceconfig2w

If you have the Windows SDK installed, you can also check the winsvc.h header,
which says: