(PUP-10857) Correctly retrieve the groups of an user
In a previous commit we mistakenly called the wrong method to get the
groups of an user, as `get` retrieves information from an user's passwd
structure, and the groups of an user are not present in that structure.

To fix this, call the superclass method, which should call
`Puppet::Util::POSIX.groups_of`.

Also reenable the `should_manage_groups` acceptance test.