(PUP-10896) Fix gid idempotency on user resource
This is a regression of #8501 which made `forcelocal` act different when
setting the `gid` parameter.

Since we were returning the GID of an user as a string, Puppet made the
assumption that the GID wasn't a number. This caused issues when setting
gid to a string value (like `gid => 'abc'`), which is allowed in Puppet.

To fix this, ensure we return the GID as integer after searching
`/etc/passwd`. If for some reason it cannot be converted, it is returned
as is (highly unlikely as non-numeric GIDs are not allowed in
`/etc/passwd`).