(PUP-10922) Downcase hostnames in fqdn_rand() function
The DNS standard does not require hostnames to be case-consistent
across different servers which are authoritative for the same
domain.

Therefore, the fqdn_rand function should not assume that the FQDN
of a given server, (which may rely on DNS lookup via getaddrinfo)
will always be returned with identical case.