(PUP-10942) Don't leak environment state when creating a new instance
Environment state for an expired environment leaked into newly created
instances, if the cached loader's `list` or `get_conf` methods had been called
prior to expiration.

This doesn't occur if `get` is called, because the environment is cached, and
the per-environment settings are cleared when `clear` or `clear_all` are called.

Clear per-environments settings prior to creating the environment instance.