(PUP-10942) Cache environments during list
When using versioned code deploys, `list` and `get` could return different
modulepaths for the same environment between the time new code was deployed and
when the environment was expired. Now the "list" method returns the cached
environment (if there is one) or the newly created environment object. In the
latter case, we cache the environment, so it can be cleared if/when clear or
clear_all are called for that environment.