(PUP-10949) Fix systemctl list-unit-files parsing
Some time ago the output of systemctl list-unit-files changed to include
an additional column. This probably happened within the last year or so,
as CentOS 8 has the old output, whereas Fedora 32 and Ubuntu 20.04 has
the new output.

To satisfy both cases, extend the regex that parses the output to allow
for an additional column, while maintaining the same constraints.