(PUP-7559) Pass correct mode when lookup SELinux default context
Originally when the file did not exist we simply passed the
mode as zero causing the file type bits not to be set which
means matchpatchcon(3) won't lookup the type based on the
correct type.

This changes so that we set the file type bits in the mode_t
passed down to the SELinux libraries based on the resource
ensure set to present, file, directory or link on the File
resource.

This fixes [1] and can be observed for SELinux file contexts
that modifies files in top directory but only for symlinks for
example this [2] httpd file context change.

The patch is backward compatible and does not break the API
for the methods in the Puppet::Util::SELinux module.

[1] https://tickets.puppetlabs.com/browse/PUP-7559
[2] https://github.com/fedora-selinux/selinux-policy-contrib/pull/366/commits/43318bf3e6415ea323952529afd4d3b49a8a040e