(PUP-10964) fix puppet_gem command on Windows
When puppet is used as a library, environment.bat
is not sourced and this leads to `PUPPET_DIR` to
not be set. Because `puppet_gem` is relying on that
to build the `gem.bat` path, it will end-up using
a non-existing path, making this provider not suitable.
This has been noticed when using the `puppet_gem`
with the package task in PE.

This commit updates the `puppet_gem` provider to
use `Gem.default_bindir` to determine the location
of the execatables. To avoid accidental usage of
`puppet_gem` provider with system ruby, a confine to
the existence of `aio_agent_version` facts has been added.