(PUP-10952) Respect `rich_data` setting without rereading env conf
This commit removes the on-write hook from the rich data setting,
instead updating environments to read the current value from settings
directly when queried, if not overwritten in the environment conf. The
env conf is read once and its rich_data setting (or lack thereof) is
cached, so we don't read the file over and over.