(PUP-11165) Enabled loading scripts from a `scripts` directory (backport to Puppet 6)
This adds a new subdirectory to the `Module` class `scripts/` which
automatically generates the functions `scripts?()`, `scripts()`, and
`script()` on the class for retrieving available scripts. This is part
of the [scripts
milestone](https://github.com/puppetlabs/bolt/issues?q=is%3Aopen+is%3Aissue+milestone%3AScripts)
for Bolt, which aims to eventually standardize on more specific file
loading from either the `files/` directory or `scripts/` directory in a
module.