(PUP-11004) Return new CA location as default if it exists
This commit adds logic to the default CA dir calculation to make it
return the new CA dir location `/etc/puppetlabs/puppetserver/ca` if the
new location has CA content, since in this case, we understand that
the CA has been migrated, and we want to be using the new location. When
the new location does not exist, we return the old default, since we
want to use the old location for new installs.