(PUP-10946) add max-files parameter for file and tidy resources
Adds a new parameter, `max-files`, to limit the number of resources
that may be created by types supporting recursion(file, tidy).

Default value for `max-files` is 0. In this case a warning is logged
in case the number of files exceeds hardcoded soft limit of 1000.

In case `max-files` is set to a value different that 0 and the number
of files is greater than `max-files`, a Puppet::Error is raised and the
resource is marked as failed.