(PUP-10999) Fix logon account setting for Windows services
Before this commit, most checks were done in `validate` and `munge` of
the `logonaccount` and `logonpassword` parameters. This proved to be too
early in case of manifests where the value of these parameters depended
on other actions (such as creating the user, changing its password or
adding the 'Logon as a service' right in the same manifest where we are
setting it as logonaccount for a service).

This commit ensures said munging and validation to happen only when
checking insyncness (munging) and when changes are required
(validation).