(maint) Relax win32-dir
Allow bundler to install win32-dir 0.7.2 that contains the ASLR fix.

This change affects bundler and the runtime dependencies for the win32
specific versions of the puppet gem. It is important to continue accepting
win32-dir 0.4.9, because that is what we ship in the puppet-runtime.

Since win32-dir 0.7.2 only supports ruby 2.4 and up, this change drops the ruby
2.3 and windows rspec combination. Ruby 2.3 went EOL March 28, 2018.

The change to ruby-prof was due to bundler pulling in a a version that wasn't
compatible with older ruby versions. So restrict that to ruby >= 2.5. It's only
needed when running the memory related benchmarks.