(PUP-10922) Add optional argument to downcase fqdn
Adds an optional argument to the `fqdn_rand` function so that the `fqdn` fact is
downcased when computing the seed. This way the function always produces the
same value if the fqdn fact only differs by case, but all of the `fqdn_rand`
arguments are the same. Note the downcase argument must be passed after the SEED
argument:

    fqdn_rand(100, 'expensive job 1', true)

The SEED argument defaults to nil, so this is also possible:

    fqdn_rand(100, nil, true)