(PUP-11046) Implement RFC2396 style escape/unescape
URI.escape/unescape was removed in Ruby 3, because it implements RFC 2396
escape/unescape with updates for RFCs 2732 and 2373, and those have been
superseded by RFC 3986.

To maintain 100% compatibility while supporting Ruby 3, implement the
escape/unescape methods in puppet[1]. In the future we can look at using
Addressable or using the newer URI parser in ruby.

[1] https://github.com/ruby/ruby/blob/v2_7_3/lib/uri/rfc2396_parser.rb#L24-L46