(PUP-10957) Pass binding's source location
If eval is called with a binding, then ruby 3 will report the source location as
["(eval)", 1], whereas ruby 2 reports the source location from the binding,
which in the test is the `functions4_spec.rb` file.

The `Binding#source_location` method only exists in ruby 2.6 and up, so
conditionally pass the source location if it's available.

See https://bugs.ruby-lang.org/issues/4352