(PUP-10957) Refactor service execution
The texecute and ucommand methods return nil if the child process returns 0 or
if failonfail is false, so the actual exit code must be retrieved via
$CHILD_STATUS, which can't be stubbed in ruby 3.

This commit changes services to call service_execute and service_command
instead, which return a ProcessOutput object containing the child's exitstatus.

In order to maintain compatibility, methods like Base#start continue to return
nil instead of ProcessOutput.