(PUP-10957) Add SortedSet runtime dependency
SortedSet was moved to a separate gem in ruby 3[1] so make it a runtime
dependency. The gem depends on the native rbtree gem, and SortedSet will prefer
that implementation (on all ruby versions).

There is ruby bug[2] where the first SortedSet doesn't sort when the rbtree gem
is present. For now call `SortedSet.setup` explicitly so the monkey patch occurs
before the first SortedSet is created.

[1] https://github.com/ruby/ruby/commit/a3db08d7b6ff119223f77e3df00b4f6deac971e2
[2] https://bugs.ruby-lang.org/issues/17841