(maint) Truncate file before overwriting it
Previously, we did not truncate the script when updating a variable.
Since the module only support setting variables to "YES" and "NO", the
only consequence was that when switching from "YES" to "NO" the last
character of the file was kept and we ended up with a blank line:

sshd_enable="YES"\n  -> sshd_enable="NO"\n\n

With the introduction of the support of more variables for the freebsd
provider, this result however in more breakage so we have to fix it.