(PUP-11074) Add catalog edge only if parent resource has a container
Managing resources with generators behaves differently with `puppet
resource` and `puppet apply`.

When running with `puppet apply`, the parent resource is always
contained by something like `Class[Main]`, so adding an edge between
that and our generated resource works.

When running with `puppet resource`, the parent resource is not
contained by anything (`container_of(parent_resource)` returns nil),
which adds an edge with a nil vertex to the catalog. This in turn causes
the run to fail with a NilClass exception when performing pre-run
checks[1] in the Transaction class, as the catalog now contains a nil
vertex.

[1] https://github.com/puppetlabs/puppet/blob/ab56d86ecb4b71298c75325462bf73376420bb40/lib/puppet/transaction.rb#L80