(PUP-11077) Pass SemVer prerelease and build as an array
Passing SemVer prerelease or build components as a list of arguments or a hash
added dots between each character due to confusion between Strings vs Arrays.
For example, SemVer(1, 0, 0, "rc0") resulted in prerelease "r.c.0". This isn't
an issue when passing the entire version as a string, eg "1.0.0-rc0".

Automatically wrapping the prerelease or build components in an array isn't a
breaking change, because it wasn't possible to do it previously. That's because
the function signature is defined to accept a `SemVerQualifier` which `Array` is
not:

    $ bx puppet apply -e 'notice(SemVer(1, 2, 3, ["rc4"], ["5"]))'
    Error: Evaluation Error: Error while evaluating a Function Call...
    rejected: parameter 'prerelease' expects a match for SemVerQualifier