(PUP-6585) Add Puppet::FileSystem.absolute?
The method doesn't use Pathname#absolute? because that method is extremely slow.
It calls File.basename on each path component breaking the path into a prefix
and base, making a copy of both. It then repeats the process for the prefix. It
also means the performance depends on the number of path components, which on
Windows can be large: `C:\Program Files\Puppet Labs\Puppet\puppet\lib\ruby\...`

The new absolute? method uses the ruby's algorithm from [1], which avoids string
copying and doesn't depend on the number of path components.