(PUP-6585) Avoid creating Pathname objects if we don't need to
Previously, the assert_path method converted its input to a Pathname object, but
then the FileImpl almost always converts the Pathname object back to a String,
resulting in unnecessary string copies.

If the input to assert_path is a String, eg `Puppet::FileSystem.basename('/')`,
then we end up making at least 2 copies of the string. One in
Pathname#initialize and another when File.basename calls Pathname#to_path to
convert it back to a String.

If the input to assert_path is a Pathname, then we end up making at least 3
copies. The extra copy occurs when Pathname#initialize calls Pathname#to_path.

This avoids all of that overhead in the common case where we're just calling
Puppet::FileSystem with a String. If called with a Pathname, then there is one
copy when calling Pathname#to_path.