(PUP-3995) Raise if a node and hostclass name conflict
Node and class scopes are both added to the topscope, so given a manifest like:

    node 'abc' { include abc }

We will add the node scope, but not the class scope, because we think it's
already been added. As a result the class is ignored, making it difficult to
debug. This also occurs when using hiera lookup/include:

    node 'abc' { lookup('classes').include }

This adds dup checks when adding a node or hostclass to the type collection,
similar to our existing checks. As a result, we will now fail compilation and
the error message will indicate where the node or hostclass was first defined
and where the redefinition was attempted.

    Error: Class 'mynode' is already defined (file: manifest.pp, line: 1); \
      cannot redefine (file: manifest.pp, line: 2) on node mynode

It also adds tests that show how a conflicting node can be specified as a
regexp, but a node as a string cannot.