(PUP-4884) Clean forge cache base directory
On Windows the default forge cache base directory is created from Dir.mktmpdir
and contains forward slashes: "C:/Users/josh/AppData/Local/Temp/...".
When unpacking a tarball we check that the entry to be unpacked has the
same prefix. The prefix is the result of File.expand_path, which also
returns a path with forward slashes. When the cache base directory is
changed to a Windows native path, e.g. C:\Temp, to work around long file
path issues, the check will fail.

This commit cleans the base_dir and the tar entry, so that the comparison
is always done using forward slashes.