(PUP-7493) Record desired mode with leading zeros in the report
When creating an event for a mode property change, the `desired_value` in the
report is obtained from the `Property#should` method, which calls `unmunge` to
convert the internal representation to its external form. The desired mode
may be an octal string without leading 0's or symbolic form (u+a). Ensure the
latter is padded with 0's so the `desired_value` matches the `previous_value`.