(PUP-9369) Raise if parsedfile providers returns nil
Custom providers like nagios subclass the `parse` method and may return nil,
causing `retrieve` to return nil. Previously that caused a hard to debug error
trying to call `each` on nil. Now we raise an error with the `target` and name
of the provider. Note the same check is made again after prefetch_hooks are
called.