(PUP-7785) Don't prematurely convert setting values
It was not possible to set a setting to a string consisting of all digits such
as `certname=000000000180`. This occurred, because the ConfigFile class
prematurely converted any value that matched /\d+/ to an integer and Ruby will
raise when given `Integer('000000000180')`. Similar problems were reported with
the `environment` setting.

The problem is that the conversion is done before we know the setting's type.
The conversion is also only done when parsing puppet.conf or CLI arguments. So
the value as stored in memory could be a string (if that was the default value),
or it could be an integer (if set in puppet.conf).

This change means settings must explicitly convert to an integer in their
"munge" method which is called during interpolation. This is already done for
autosign, duration, integer, port, priority and ttl settings.