(PUP-11082) Use built-in OpenSSL::PKey.read
This commit switches our code that reads private keys to use Ruby's
built-in `PKey.read` method, instead of constructing `PKey` objects
directly.

Previously, we couldn't use this function because it was broken in
older versions of Ruby, and the code that calls it was used in JRuby,
where `PKey.read` is still broken. But we no longer load private keys
in JRuby, and in Puppet 7 we no longer support the older versions of
Ruby where `PKey.read` was broken. So it is now safe to use the more
robust API method.

This enables us to correctly read elliptic curve keys in PKCS#8 format,
which is now the standard. It uses the "BEGIN PRIVATE KEY" header,
which is also readable by the Java code on the server. Previously we
required EC keys to begin with "BEGIN EC PRIVATE KEY", which is not
supported by Bouncy Castle.

Because JRuby does not support `PKey.read`, this commit also skips
running some additional tests that no longer work correctly when run in
JRuby.

(PUP-11082) Skip private key loading tests on JRuby