(PUP-11079) Tidy files with age equal or greater
Tidy resource documentation for the age parameter states that files
whose age is equal or greater than the specified time will be removed,
and that setting age to 0 will remove all files.[1]

The actual implementation only removes the file if the age is greater,
not equal, so fix the comparison.

Note that the size parameter is documented similarly, and the
implementation is correct there.

[1] https://puppet.com/docs/puppet/7.7/types/tidy.html#tidy-attribute-age