(PUP-6801) Honor agent section for agent_specified_environment fact
When running `facter -p`, if the environment is set in the agent section
of puppet.conf, it would not be reported in the
agent_specified_environment fact due to Puppet running in the user run
mode.

Now an additional check is made to see if the setting is explicitly set
in the agent section, and the fact prioritizes the agent section over
other sections.