(PUP-5704) Let exec control shell invocation
The posix implementation of the exec provider relies
on `Puppet::Util::Execution` which passes the command
string to `Kernel.exec`[0]. By default, the command
line is passed to the standard shell. To avoid this,
a new parameter was added `pass_through_shell`,
which defaults to `true` and can be set to false.
The command is split as `cmdname, arg1, ...`
before passing it to `exec`.

[0] https://www.rubydoc.info/stdlib/core/Kernel:exec