(PUP-10967) Debug logs clean up when resolving account SID on Windows
This commit restricts the `Could not retrieve raw SID bytes from` debug
log to when an actual SID is passed to `ConvertStringSidToSidW`. The
order in which the `Puppet::Util::Windows::SID.name_to_principal` method
checks and tries to convert the input (be it an account name, domain
qualified account name or SID) to a
`Puppet::Util::Windows::SID::Principal` object made it print error
messages to debug when it was actually succeeding in the end. Since
every Puppet Agent run needs information about at least the currently
logged on user, this was creating unnecessary noise in the debug logs.